excluding the triton kernels from the code coverage, not handled properly
