+++ .coveragerc
@@ -4,3 +4,4 @@ omit =
     tests/*
     setup.py
     xformers/benchmarks/*
+    xformers/triton/k_*
